LayoutTests/media/video-poster.html

<video poster="content/greenbox.png" controls></video>
<script src=video-test.js></script>
<script>
    testExpected("video.getAttribute('poster')", "content/greenbox.png");
    testExpected("relativeURL(video.poster)", "content/greenbox.png");
    video.setAttribute('poster', 'content/abe.png');
    testExpected("video.getAttribute('poster')", "content/abe.png");
    testExpected("relativeURL(video.poster)", "content/abe.png");
    endTest();
</script>
