LayoutTests/media/video-poster-expected.txt


EXPECTED (video.getAttribute('poster') == 'content/greenbox.png') OK
EXPECTED (relativeURL(video.poster) == 'content/greenbox.png') OK
EXPECTED (video.getAttribute('poster') == 'content/abe.png') OK
EXPECTED (relativeURL(video.poster) == 'content/abe.png') OK
END OF TEST
